remove discord override and update

+++ src/discord.hs
@@ -102,5 +102,5 @@ main = void . runExceptT $ (do
     token <- catchE (ExceptT $ tryJust (guard . isDoesNotExistError) (getEnv "DISCORD_AUTH_TOKEN")) . const $ fail "Failed to get the authentication token. Please set the environment variable DISCORD_AUTH_TOKEN to your token & make sure you include DISCORD_CHANNEL_ID. See https://github.com/aquarial/discord-haskell/wiki/Creating-your-first-Bot for more details."
     channelId <- catchE (ExceptT $ tryJust (guard . isDoesNotExistError) (getEnv "DISCORD_CHANNEL_ID")) . const $ fail "Failed to get the channel ID. Please set the environment variable DISCORD_CHANNEL_ID."
     putStrLn "Starting bot"
-    void . liftIO . runDiscord . runDiscordOpts (T.pack token) $ fromIntegral (read channelId)
+    void . liftIO . runDiscord . runDiscordOpts (T.pack token) $ read channelId
     putStrLn "Bot stopped")
